% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.compute.max.R
\name{spss.compute.max}
\alias{spss.compute.max}
\title{Generate SPSS 'COMPUTE' syntax to compute the MAX() of two (sets of) variables.}
\usage{
spss.compute.max(var1, var2, append = "new.", ...)
}
\arguments{
\item{var1}{the first argument for MAX(). Used for the naming of the output variable(s).}

\item{var2}{the second argument for MAX().}

\item{append}{specifies the text that should be appended to the name(s) of the variable(s) in var1 to create the output variable(s).}

\item{...}{any additional arguments that can be passed to functions spss.format.max and spss.format.compute}
}
\description{
Generate SPSS 'COMPUTE' syntax to compute the MAX() of two (sets of) variables.
}
\examples{
spss.compute.max(c("dob", "income"), c("dob_f", "income_f"), append = "total.")
spss.compute.max(c("dob", "income"), c("dob_f", "income_f"))
}
