% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.concat.new.R
\name{spss.concat.new}
\alias{spss.concat.new}
\title{Generate SPSS 'STRING' syntax and 'COMPUTE' syntax to compute the CONCAT() of two (sets of) variables.}
\usage{
spss.concat.new(str1, str2, file = FALSE, ext = "txt", append = "",
  name = "", ...)
}
\arguments{
\item{str1}{the first argument for CONCAT(). Used for the naming of the output variable(s).}

\item{str2}{the second argument for CONCAT().}

\item{file}{specifies if the function should output a file. Defaults to FALSE.}

\item{ext}{specifies the desired extension of the output file. Defaults to "txt".}

\item{append}{specifies the text that should be appended to the name(s) of the variable(s) in str1 to create the output variable(s). Defaults to an empty string.}

\item{name}{specifies the name of the new String variable to be created. Defaults to an empty string ; if left unspecified, the function will use the append parameter.}

\item{...}{any additional arguments that can be passed to functions spss.string and spss.compute.concat}
}
\description{
Generate SPSS 'STRING' syntax and 'COMPUTE' syntax to compute the CONCAT() of two (sets of) variables.
}
\examples{
spss.concat.new(c("dob", "income"), c("dob_f", "income_f"), append = "total.")
spss.concat.new(c("dob", "income"), c("dob_f", "income_f"), file = TRUE, ext = "sps")
}
