% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.format.compute.R
\name{spss.format.compute}
\alias{spss.format.compute}
\title{Generate SPSS 'COMPUTE' syntax to carry out an operation on two (sets of) variables.}
\usage{
spss.format.compute(output, operation)
}
\arguments{
\item{output}{specifies the name(s) of the output variable(s) of COMPUTE().}

\item{operation}{specifies the computational operation to be carried out.}
}
\description{
Generate SPSS 'COMPUTE' syntax to carry out an operation on two (sets of) variables.
}
\examples{
spss.format.compute(new.var, "MAX(var1,var2)")
spss.format.compute(c("dob", "comments"), c("MAX(dob,dob_f)", "CONCAT(comments,comments_f)"))
}
