% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.format.concat.R
\name{spss.format.concat}
\alias{spss.format.concat}
\title{Generate SPSS 'concat()' syntax for two (sets of) variables.}
\usage{
spss.format.concat(var1, var2)
}
\arguments{
\item{var1}{the first argument for concat().}

\item{var2}{the second argument for concat().}
}
\description{
Generate SPSS 'concat()' syntax for two (sets of) variables.
}
\examples{
spss.format.concat(c("dob", "income"), c("dob_f", "income_f"))
spss.format.concat("income", "income_f")
}
