% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.format.max.R
\name{spss.format.max}
\alias{spss.format.max}
\title{Generate SPSS 'MAX()' syntax for two (sets of) variables.}
\usage{
spss.format.max(var1, var2)
}
\arguments{
\item{var1}{the first argument for MAX().}

\item{var2}{the second argument for MAX().}
}
\description{
Generate SPSS 'MAX()' syntax for two (sets of) variables.
}
\examples{
spss.format.max(c("dob", "income"), c("dob_f", "income_f"))
spss.format.max("income", "income_f")
}
