% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.rtrim.R
\name{spss.rtrim}
\alias{spss.rtrim}
\title{Generate SPSS 'RTRIM' syntax to apply RTRIM() to a string variable.}
\usage{
spss.rtrim(str)
}
\arguments{
\item{str}{the string argument for RTRIM().}
}
\description{
Generate SPSS 'RTRIM' syntax to apply RTRIM() to a string variable.
}
\examples{
spss.rtrim("variable_to_be_trimmed")
}
