% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.string.R
\name{spss.string}
\alias{spss.string}
\title{Generate SPSS 'STRING' syntax to create a variable of type string.}
\usage{
spss.string(names, string.format = "A15")
}
\arguments{
\item{names}{the name(s) of the variable(s) to be created.}

\item{string.format}{specifies the formatting to use when creating the string variable. Defaults to "A15".}
}
\description{
Generate SPSS 'STRING' syntax to create a variable of type string.
}
\examples{
spss.string("my.string")
spss.string("long.string", string.format = "A40")
}
