% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.compute.concat.R
\name{spss.compute.concat}
\alias{spss.compute.concat}
\title{Generate SPSS 'COMPUTE' syntax to compute the CONCAT() of two (sets of) variables.}
\usage{
spss.compute.concat(str1, str2, names)
}
\arguments{
\item{str1}{the first argument for CONCAT().}

\item{str2}{the second argument for CONCAT().}

\item{names}{the name(s) of the variable(s) to be created.}
}
\description{
Generate SPSS 'COMPUTE' syntax to compute the CONCAT() of two (sets of) variables.
}
\examples{
spss.compute.concat("feedback", "feedback_f", "enfr.feedback")
spss.compute.concat("feedback", "feedback_f", "enfr.feedback")
}
