% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.compute.max.R
\name{spss.compute.max}
\alias{spss.compute.max}
\title{Generate SPSS 'COMPUTE' syntax to compute the MAX() of two (sets of) variables.}
\usage{
spss.compute.max(var1, var2, output = FALSE, ext = "txt",
  append = "new.")
}
\arguments{
\item{var1}{the first argument for MAX().}

\item{var2}{the second argument for MAX(). Used for the naming of the output variable(s).}

\item{output}{specifies if the function should output a file. Defaults to FALSE.}

\item{ext}{specifies the desired extension of the output file. Defaults to "txt".}

\item{append}{specifies the text that should be appended to the name(s) of the variable(s) in var2 to create the output variable(s).}
}
\description{
Generate SPSS 'COMPUTE' syntax to compute the MAX() of two (sets of) variables.
}
\examples{
spss.compute.max(var1, var2, append = "total.")
spss.compute.max(var1, name1, output = TRUE, ext = "sps")
}
