% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.rename.R
\name{spss.rename}
\alias{spss.rename}
\title{Generate SPSS 'RENAME' syntax to rename a (set of) variables into another.}
\usage{
spss.rename(values, rename)
}
\arguments{
\item{values}{the variable(s) to be renamed.}

\item{rename}{the name(s) to use for renaming.}
}
\description{
Generate SPSS 'RENAME' syntax to rename a (set of) variables into another.
}
\examples{
spss.rename(c("oldname1", "oldname2", "oldname3"), c("new1", "new2", "new3"))
spss.rename(c("oldname1", "oldname2"), c("new1", "new2"))
}
