% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spss.rename.R
\name{spss.rename}
\alias{spss.rename}
\title{Generate SPSS 'RENAME' syntax to rename a (set of) variables into another.}
\usage{
spss.rename(values, rename, output = FALSE, ext = "txt")
}
\arguments{
\item{values}{the variable(s) to be renamed.}

\item{rename}{the name(s) to use for renaming.}

\item{output}{specifies if the function should output a file. Defaults to FALSE.}

\item{ext}{specifies the desired extension of the output file. Defaults to "txt".}
}
\description{
Generate SPSS 'RENAME' syntax to rename a (set of) variables into another.
}
\examples{
spss.rename(var1, name1)
spss.rename(var2, name2, output = TRUE, ext = "sps")
}
